requirement.txt
customtkinter
pillow
CTkMessagebox
tkcalendar
pywin32